Moved files to root and removing text area logic
