+++ src/components/CustomNodes/TriangleNode.jsx
